Get places in batches in overview
This fetches places in batches of 2000 from the API, while only allowing 3 running requests at the same time, fixing timeouts on the production servers.

When a new batch of places is arriving, it’s directly clustered with leaflet (a CPU-expensive op). This way, clustering runs in parallel with the requests.