Reduce debug output for skipped PoI on import
When a source has many skipped PoIs, each skipped PoI would generate an update MongoDB statement, leading to a long queue of MongoDB operations that would all be published to connected clients that subscribed the source.

This slows down the server while a source stream chain is being implemented because typically, in the first steps, the `originalId` of a PoI is missing.

With the change, only one PoI with a missing `originalId` attribute is output for debugging.